clean up types, rename methods for long term
