+++ great_expectations/rule_based_profiler/__init__.py
@@ -1,1 +1,1 @@
-from .profiler import RuleBasedProfiler
+from .rule_based_profiler import RuleBasedProfiler
