+++ tests/rule_based_profiler/alpha/test_profiler_user_workflows.py
@@ -12,7 +12,7 @@ from great_expectations import DataContext
 from great_expectations.core import ExpectationSuite
 from great_expectations.core.batch import BatchRequest
 from great_expectations.datasource import DataConnector, Datasource
-from great_expectations.rule_based_profiler.profiler import RuleBasedProfiler
+from great_expectations.rule_based_profiler.rule_based_profiler import RuleBasedProfiler
 from great_expectations.validator.metric_configuration import MetricConfiguration
 from great_expectations.validator.validator import Validator
 
@@ -91,7 +91,7 @@ def test_alice_profiler_user_workflow_single_batch(
         data_context=data_context,
     )
 
-    expectation_suite: ExpectationSuite = profiler.profile(
+    expectation_suite: ExpectationSuite = profiler.run_profiler(
         expectation_suite_name=alice_columnar_table_single_batch[
             "expected_expectation_suite_name"
         ],
@@ -104,6 +104,7 @@ def test_alice_profiler_user_workflow_single_batch(
     )
 
 
+# noinspection PyUnusedLocal
 def test_bobby_columnar_table_multi_batch_batches_are_accessible(
     monkeypatch,
     bobby_columnar_table_multi_batch_deterministic_data_context,
@@ -200,7 +201,7 @@ def test_bobby_profiler_user_workflow_multi_batch_row_count_range_rule_and_colum
         data_context=data_context,
     )
 
-    expectation_suite: ExpectationSuite = profiler.profile(
+    expectation_suite: ExpectationSuite = profiler.run_profiler(
         expectation_suite_name=bobby_columnar_table_multi_batch[
             "test_configuration_oneshot_sampling_method"
         ]["expectation_suite_name"],
@@ -244,7 +245,7 @@ def test_bobster_profiler_user_workflow_multi_batch_row_count_range_rule_bootstr
         data_context=data_context,
     )
 
-    expectation_suite: ExpectationSuite = profiler.profile(
+    expectation_suite: ExpectationSuite = profiler.run_profiler(
         expectation_suite_name=bobster_columnar_table_multi_batch_normal_mean_5000_stdev_1000[
             "test_configuration_bootstrap_sampling_method"
         ][
